Update Fix Image Upload

+++ src/app/api/[admin]/settings/upload/route.ts
@@ -44,26 +44,49 @@ export async function POST(req: Request) {
     }
 
     const ext = path.extname(file.name);
-    const fileName = `${Date.now()}-${Math.random().toString(36).substring(2)}${ext}`;
-    const uploadDir = path.join(process.cwd(), 'public', 'uploads', 'settings');
-    const relativePath = `/uploads/settings/${fileName}`;
+    const filename = `${Date.now()}-${Math.random().toString(36).substring(2)}${ext}`;
 
-    await mkdir(uploadDir, { recursive: true });
+    if (process.env.APP_ENV === 'local') {
+      const uploadDir = path.join(process.cwd(), 'public', `${process.env.UPLOAD_PATH}`, 'settings');
+      await mkdir(uploadDir, { recursive: true });
 
-    await writeFile(path.join(uploadDir, fileName), buffer);
+      const filePath = path.join(uploadDir, filename);
+      await writeFile(filePath, buffer);
 
-    if (oldFile) {
-      const oldPath = path.join(process.cwd(), "public", oldFile);
-      try {
-        await unlink(oldPath);
-      } catch {
-        logger.warn("ลบไฟล์เก่าไม่สำเร็จ (อาจไม่มีอยู่แล้ว)");
+      if (oldFile) {
+        const oldPath = path.join(process.cwd(), "public", oldFile);
+        console.log(process.env.APP_ENV + ' : ' + oldPath)
+        try {
+          await unlink(oldPath);
+        } catch {
+          logger.warn("ลบไฟล์เก่าไม่สำเร็จ (อาจไม่มีอยู่แล้ว)");
+        }
       }
+    } else {
+      const uploadPath = process.env.UPLOAD_PATH;
+      if (uploadPath) {
+          const filePath = path.join(uploadPath, 'settings', filename);
+          await writeFile(filePath, buffer);
+
+          if (oldFile) {
+            const oldPath = path.join(uploadPath, oldFile);
+            logger.debug(process.env.APP_ENV + ' : ' + oldPath)
+            try {
+              await unlink(oldPath);
+            } catch {
+              logger.warn("ลบไฟล์เก่าไม่สำเร็จ (อาจไม่มีอยู่แล้ว)");
+            }
+          }
+      } else {
+          return NextResponse.json({ message: 'ไม่สามารถอัปโหลดไฟล์ได้!' }, { status: 500 });
+      }
+      
     }
 
-    return NextResponse.json({
-      url: relativePath
-    });
+    const imageUrl = `${process.env.APP_ENV !== 'local' ? process.env.CDN_URL : process.env.UPLOAD_PATH}/settings/${filename}`;
+
+    return NextResponse.json({ message: 'File uploaded successfully.', url: imageUrl }, { status: 201 });
+
   } catch (err) {
     logger.error("UPLOAD ERROR:", { error: err });
     return NextResponse.json({ error: "Upload failed" }, { status: 500 });
